+++ components/vf-button/CHANGELOG.md
@@ -1,3 +1,7 @@
+### 3.0.3
+
+* Changed: Test cases added for React  [Tracking issue](https://github.com/visual-framework/vf-core/issues/2174)
+
 ### 3.0.2
 
 * Changed: Test cases added for Angular  [Tracking issue](https://github.com/visual-framework/vf-core/issues/1932)